pyproject.toml
[tool.poetry]
name = "checkbox"
version = "0.1.0"
description = ""
authors = ["Dmytro Podolynnyi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
virtualenv = "20.25.3"
fastapi = "^0.115.8"
SQLAlchemy = "^2.0.31"
alembic = "^1.13.2"
pydantic = "^1.9.1"
asyncpg = "^0.29.0"
uvicorn = "0.30.1"
gunicorn = "^20.1.0"
black = "^24.8.0"
Jinja2 = "^3.1.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "4.0.1"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pytest-mock = "^3.6.1"
pytest-cov = "^3.0.0"
pre-commit = "^2.16.0"
python-dotenv = "^0.20.0"
factory-boy = "^3.2.1"
psycopg2-binary = "^2.9.5"
faker = "18.3.4"
pytest-socket = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
